chore: remove unused, unstable 'http' namespace
