+++ runtime/shared.rs
@@ -45,7 +45,6 @@ extension!(runtime,
     "13_buffer.js",
     "30_os.js",
     "40_fs_events.js",
-    "40_http.js",
     "40_process.js",
     "40_signals.js",
     "40_tty.js",
