+++ tests/testdata/run/unstable_http.disabled.out
@@ -1,12 +1,4 @@
 main undefined
 main undefined
-main undefined
-main undefined
-main undefined
-main undefined
-worker undefined
-worker undefined
-worker undefined
-worker undefined
 worker undefined
 worker undefined
