+++ tests/testdata/run/unstable_http.enabled.out
@@ -1,12 +1,4 @@
 main [class HttpClient]
 main [Function: createHttpClient]
-main [class HttpConn]
-main Symbol("[[associated_ws]]")
-main [Function: serve]
-main [Function: upgradeWebSocket]
 worker [class HttpClient]
 worker [Function: createHttpClient]
-worker [class HttpConn]
-worker Symbol("[[associated_ws]]")
-worker [Function: serve]
-worker [Function: upgradeWebSocket]
