+++ tools/core_import_map.json
@@ -237,7 +237,6 @@
     "ext:runtime/13_buffer.js": "../runtime/js/13_buffer.js",
     "ext:runtime/30_os.js": "../runtime/js/30_os.js",
     "ext:runtime/40_fs_events.js": "../runtime/js/40_fs_events.js",
-    "ext:runtime/40_http.js": "../runtime/js/40_http.js",
     "ext:runtime/40_process.js": "../runtime/js/40_process.js",
     "ext:runtime/40_signals.js": "../runtime/js/40_signals.js",
     "ext:runtime/40_tty.js": "../runtime/js/40_tty.js",
